More relevant return code (#1147)
* Return the adequate error code

Signed-off-by: Gaël Écorchard <gael@km-robotics.cz>

* Return more appropriate status code in AnytimePathShortening

The AnytimePathShortening planner now returns INVALID_START or
INVALID_GOAL if any planner returned such status. Not all planners are
able to return INVALID_START or INVALID_GOAL, so this status is returned
if at least one planner returned it. If no planner returned
INVALID_START or INVALID_GOAL, the status is EXACT_SOLUTION or UNKNOWN.

Signed-off-by: Gaël Écorchard <gael@km-robotics.cz>

* Render code more readable and consistent

Signed-off-by: Gaël Écorchard <gael@km-robotics.cz>

---------

Signed-off-by: Gaël Écorchard <gael@km-robotics.cz>
Co-authored-by: Gaël Écorchard <gael@km-robotics.cz>